x11-libs/fltk: don't install fluid and fltk-options twice
Just noticed tha had both `/usr/bin/fluid` and `/usr/bin/fluid-shared`
and the .desktop points only to `fluid`, likewise for fltk-options.
Not seeing a point in the static version on Gentoo (it is also several
times bigger), so overwrite them with the -shared copy.

Not revbumping given this is just a minor space saving still and
not worth a rebuild.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>