app-misc/anki: ensure GUI was built before compiling Sphinx docs
The GUI build generates certains directories listed in
python/sphinx/conf.py, which Sphinx AutoAPI documents or errors out if
these are not present. We rely on the ordering of the sub-phase
functions python_compile -> python_compile_all to run the GUI build and
Sphinx in the correct order.

Since the Sphinx output documents the Python modules used in the GUI,
providing a workaround for USE="doc,-gui" makes no sense.

Fixes: e6b81091d520c123d552d521073088a19748ee78
Signed-off-by: Lucio Sauer <watermanpaint@posteo.net>