media-video/ffmpeg: sync comment for 4.4.x
Just to have less differences between ebuilds where unneeded.

This was slightly reworded given >6 now handles some flags in
multilib_src_configure, and this is to make it sound it's primarily
for the lto bit and not all flags (...not that it matters in 4.4.x
given it disables lto entirely).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>