profiles: package.mask current nvidia-drivers-580.x versions
Was already planning to not stabilize these versions, but IMO has
too many problems even for ~testing users without explicit opt-in.

NVIDIA is aware and next version should hopefully be better, so
limiting the masks to the current versions.

As noted in the mask message, feel free to unmask if been using it
without problems (does not necessarily affect everyone).

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>