media-video/ffmpeg: add 8.0_pre20250812 (unkeyworded for testing)
Given some testing has started, it'll be simpler to have this for
reference. Does not restore libpostproc which may or may not be part
of the final release tarball.

On a side-note, can also use 9999 with the 8.0 branch by doing:

	EGIT_BRANCH=release/8.0 emerge =ffmpeg-9999

Albeit this may break in the future if ebuild receives breaking
updates (no plans to introduce ffmpeg-8.0.9999 in ::gentoo for
this currently).

Bug: https://bugs.gentoo.org/961435
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>