dev-python/pygame: skip likely flaky test_queue
Fails for me, and according to an upstream issue that test depends
on timing and may be flaky.

Albeit alternatively it *could* be related to newer libsdl2/sdl2-mixer
but I haven't spent more time on this. I do not believe this test
failing will necessarily imply runtime issues either way (upstream also
skips it for one platform).

Closes: https://bugs.gentoo.org/951940
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>