dev-python/pygame: enable pypy3_11
Came here to see if 3.14 could be added given pygame is blocking adding
3.14 to qutebrowser in a long chain of test dependencies, but that
failed.

There is a recent PR upstream to add 3.14, but it has failing CI and
nothing to backport -- figure will leave this alone for now and see how
that progress.

On the other hand, tested pypy3_11 at same time and it passed -- so may
as well get it out of the way wrt bug #951020. Unsure if we really want
to keep pypy support on this package but adding given the bug exists
and it had (no-op) pypy3 in COMPAT.

Did have a bit of problems to get tests to pass at all for *any* python
version though, not 100% sure what was breaking them but it worked
after some messing around (may have been related to me using
unkeyworded libsdl2-2.32.56 which actually uses libsdl3, but not going
to worry about that right now).

Closes: https://bugs.gentoo.org/951020
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>