app-shells/bash: move the skeleton files to ${FILESDIR}/skel
The top of FILESDIR has become a little unwieldy. In advance of
submitting a patch that will introduce yet more files to be tracked, I
thought that it would be a good opportunity to conduct a minor
re-organisation. To that end, this commit relocates the skeleton files
to a sub-directory.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>