sys-apps/locale-gen: add 3.2-r1 and 3.3-r1 and revise 9999 per bug #961743
Upon locale-gen-3.0 being introduced, the ebuild lost the ability to
prefixify the locale-gen script in src_prepare(). Address this issue so
that the shebang is properly adjusted for Gentoo Prefix environments.

I eschewed the use of the prefix eclass because I consider its
implementation to be substandard. I also eschewed the use of sed(1)
because it encourages pattern injection and seldom returns a useful exit
status. Instead, perl is used to replace the shebang. I could think of
no reason not to, given that dev-lang/perl is already a requirement.

Reported-by: Nick Bowler <nbowler@draconx.ca>
Fixes: ecc3beb504587ee532156b4f3f1c82edfd6bf87f
Closes: https://bugs.gentoo.org/961743
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>