profiles/targets/desktop: Add pulseadio
To match both the GNOME and Plasma profiles this will add pulseaudio
to the normal dektop profile to give WM and the other DE's the same
audio experience.

This will make it easy for users to switch to Pipewire as a bonus
and allow people that wish to use ALSA only to just set -pulseaudio
in their make.conf.

Closes: https://bugs.gentoo.org/961764
Signed-off-by: Ian Jordan <immoloism@gmail.com>