dev-util/debugedit: make dwz an optfeature
mgorny and I discussed this ages ago but I never implemented it. dwz
is only used for `find-debuginfo` which basically nobody is going to use
in Gentoo (at least right now, see bug #938763 and bug #947387).

Avoid pulling in dwz on the system of everybody who wants installsources. This
is especially important as I'm contemplating proposing having debugedit in
@system.

Bug: https://bugs.gentoo.org/938763
Bug: https://bugs.gentoo.org/947387
Signed-off-by: Sam James <sam@gentoo.org>