media-video/ffmpeg: add 8.0 (unkeyworded)
While issues seem generally trivial to fix, believe there is still
a decent amount of broken packages, so leaving it unkeyworded for now.

libpostproc isn't included in the tarball as-is after all. It *would*
be possible to restore it as a "source plugin" if really had to, but
sounds messy as it uses a script that seems to pull from libpostproc
git master and would rather not have to deal with this if possible.
Could revisit before rekeywording if we really need it.

Bug: https://bugs.gentoo.org/961435
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>