dev-util/maturin: tentatively stop skipping macosx test on sparc
Been over a year, would believe it's likely fine by now but it hasn't
been tested (odds are it may not be often given sparc is due to be
destabilized soon, maturin hasn't gotten a sparc stable since 1.8.6).

Feel free to re-open or open a new bug if actually run into this
again to confirm it's not obsolete, then will skip it again fwiw.

Closes: https://bugs.gentoo.org/934573 (obsolete)
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>