dev-python/qtpy: remove heavily outdated sed
```
 * SED: the following did not cause any changes
 *     sed -i -e 's:--cov=qtpy --cov-report=term-missing::' pytest.ini || die;
 * no-op: -e s:--cov=qtpy --cov-report=term-missing::
```

In commit d27e3d368dbe284d91ccb28fced0293fbc6e4be5, this package added
`-o addopts=` to the pytest invocation, wiping out the effects of this
line that was being sedded. Subsequently, the fact that the sed failed
to apply due to incidental refactoring, went unnoticed. The problematic
`--cov*` arguments are still there, and could be deleted, but is now
redundant.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>