dev-lang/ngn-k: Add live ebuild support for the ngn/k derivative of APL.
Add support for the k derivative of APL. This is important because gentoo currently does not have any version of APL or its derivatives, and APL is an important language. ngn/k is a stable version of k that is minimal, easy to build and not actively modified. However, it lacks a release or versioning history. Thanks to APL orchard for suggesting ngn/k for gentoo.

Suggested-by: pmoriarty, et al. /c.f./ https://news.ycombinator.com/item?id=21652218
Signed-off-by: Sylvian Hemus <shemus@sylvianhemus.com>