media-sound/ncspot: drop 1.2.2, 1.3.0
This drops the stable versions. This is one of the projects where the previous
release is broken due to changing API by Spotify, and therefore any updates
should be pushed straight to stable or fast-stabilized immediately after
pushing. Since I don't think this is a particularly popular application,
dropping it back to ~testing seems like the way to go for now.

Signed-off-by: Joonas Niilola <juippis@gentoo.org>