dev-libs/md4c: keyword 0.5.2 for ~loong
Untested, blindly adding (timeout) because dropping ~loong from all of
dev-qt/*-6.9.2 just because of this would be a pain and do not really
want to delay the bump either.

Thankfully md4c is rather simple, and sounds unlikely to be broken.
Not to mention that qtbase used to use a bundled copy of md4c and it
worked fine on ~loong.

Closes: https://bugs.gentoo.org/957680
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>