sci-misc/vitables: add 3.1.0, drop 3.0.2_p4-r1
Drastically updated version on account of it includes loads of patches
we were importing from Debian (?) and also actually works with numpy
2.x. In fact, its ::gentoo deps already require `>=numpy-2` therefore it
has so far been impossible to install it in working at all. As such, I
give no credence to the idea of keeping the old version around "in case
people encounter issues with the new release". Not this time, buddy!

In the process, it turns out a particular test is broken. It tries to
test locale support, which would be nice and all except there are some
raw *.ts files floating around in the repo root and no way to actually
compile or install them. To boot, the testcase string wants to hardcode
the es_ES locale. There is no upstream CI or anything.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>