games-emulation/higan: drop gtk USE (always enabled)
higan is mostly in low maintenance mode with the author gone and kind
of obsolete with all the alternatives, so porting efforts are likely
to take a long time if ever (not to mention it is buggy with Qt5
as-is). May as well stick to the less problematic options bare
straight up last-riting.

Has a not-in-tree fork (ares) but it does not support Qt currently,
so either way Qt likely has to go.

Closes: https://bugs.gentoo.org/962040
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>