dev-python/sphinx: allow newer docutils
Upstream pins to max bounds, due IIRC to tight coupling with format
specifics of docutils output. We just stabled a newer docutils, and it
is incompatible with sphinx RDEPEND.

Upstream allows the newer version in git only. There are no
corresponding changes other than an earlier testfix. Backporting that is
sufficient to pass all tests with a loosened RDEPEND.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>