app-emulation/libguestfs: add 1.56.2
The requirement for app-admin/augeas and app-misc/hivex[ocaml?]
was removed if the daemon is not being build (which is
unsupported in Gentoo)
Dropped use of dev-ml/ounit2[ocamlopt] for tests.
Does not use flex or bison, so dropped in BDEPEND.
Upstream Changed from jansson to c-json as their JSON library.
Some previopusly failing tests were re-enabled
Intospection support is broken and officially depreceated. Upstream will
strip it out completely in 1.58.0, so therefore removed introspection
(and Vala) support.

Signed-off-by: Christopher Byrne <salah.coronya@gmail.com>