dev-util/maturin: avoid cargo_src_install
Eli pointed out that maturin was (partly) being recompiled during
src_install for nothing, aka it sees that "something" changed and
thinks it needs to recompile some bits. I'd imagine that this
something comes from the python eclasses (CPPFLAGS change maybe?),
but I haven't dug into it as it's simpler+faster to just use dobin
here, other bits are handled by setuptools and then cargo_src_install
does nothing else (because of that, I see no real reason to revert
this even if it the recompiling stopped -- it's faster to skip cargo's
checks and we have no real "files to install" to track).

Some bits may get recompiled during src_test as well alongside tests
themselves but not going to worry too much when it's just for tests
and doesn't hit common users.

Thanks-to: Eli Schwartz <eschwartz@gentoo.org>
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>