media-libs/libplacebo: unkeyword 7.351.0 for ~hppa
vulkan is not available on hppa, and it's generally kind of weird
to use libplacebo without vulkan even if it's possible (lot of features
are unavailable, and e.g. ffmpeg requires libplacebo[vulkan]).

This was formerly keyworded (only) for mpv, but mpv lost that keyword
due to failing tests (bug #921091) combined with low interest a while
ago.

wrt masks, had formerly restored the vulkan mask but misunderstood the
reason for removing it (it was fine, could debate keeping it in the event
that vulkan is globally unmasked but that won't happen on hppa). We can
instead just clean that up now -- odds are we won't need it again.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>