net-misc/kea: Stop old kea service on upgrade to >= 3.0
Add pkg_preinst to stop previous kea service, if we are using openrc and
the previous version is < 1.3. Unfortunately there is no simple and fool
proof way to re-instate the services using the new init skeleton without user
intervention and if we don't stop the old service first the user
will find it hard to do post install.

Signed-off-by: Peter Leese <inbox@peterleese.org>
Part-of: https://github.com/gentoo/gentoo/pull/43614
Signed-off-by: Sam James <sam@gentoo.org>