python-utils-r1.eclass: replace 2to3 with "unsupported" wrapper for 3.13+
It has been removed upstream. Due to the use of a symlink for wrapping
it (as opposed to a shell script like python / python-config) this meant
searching PATH didn't find the broken symlink, and instead found
/usr/bin/2to3-3.12 with a 3.13 impl.

Check the version of EPYTHON and add it to nonsupp wrappers instead of
making a symlink.

Reported-by: Paul Zander <negril.nx+gentoo@gmail.com>
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>
Signed-off-by: Michał Górny <mgorny@gentoo.org>