dev-libs/jerasure: don't skip test failing w/ F_S_=3
encode_decode.sh fails w/ _FORTIFY_SOURCE=3 but the test is new in the
snapshot we're taking (there were no tests before), and upstream development
is over. I originally skipped the test (or, well, XFAIL'd it), but that
doesn't work when the default is stil _F_S=2 and we only default to _F_S=3
on hardened profiles.

So, let's repurpose the bug toralf failed for the XPASS by dropping the
XFAIL and use the bug to track it failing w/ _F_S=3.

Bug: https://bugs.gentoo.org/960946
Signed-off-by: Sam James <sam@gentoo.org>