linux-mod-r1.eclass: silence spurious warning with _p* kernels
Like noted in the comment, ideally would want to check for the
actual matching version but we'd need to test+sanitize KV_EXTRA
and figure out if it's really a version component as it could be
anything.

_p being used should be a rare event and it shouldn't be a big
deal if the check ignores it.

Skipping ML review given trivial and is just about a warning.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>