dev-qt/qttools: update linguist description for 6.10
Newly checks for Quick to be available when want the GUI linguist.
Not really worth setting up per-version restrictions.

The handful of users using GUI linguist with USE=-qml may get
surprised that it's gone. Feel it's not worth adding a warning
given desktop profiles enable qml by default and developers
using Qt will typically enable it if not already.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>