net-misc/wakeonlan: revbump to 0.42
And dropping the downstream patch to read hostnames from /etc/ethers.
That patch should be upstreamed, or users should maybe wrap their own
scripts around given the tool officially lacks the feature.

We add an ewarn to tell users about the removed feature. All tests are
currently disabled and we run an empty suite on src_test. This way we
can not miss possible functional tests in future versions.

Signed-off-by: Henning Schild <henning@hennsch.de>