dev-qt/qtwebengine: update live
Upon closer look, openh264 seems unused with our current configuration
(both system's and bundled from a quick grep) and this just adds a
dependency or potential automagic for nothing.

Originally I thought it could be used with USE=bindist (no real reason
not to if system's), but it has configure.cmake checks to specifically
disable it when proprietary codecs are not enabled. Either way, the
code seems only for webrtc and wouldn't offer h264 playback with
bindist so the interest here was mostly to unbundle than functionality.

That aside, also noticed that some test junk was newly being installed,
(it's a whole QML libary so kind of annoying), and also that disk space
usage went over 10GB.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>