sys-auth/sssd: Allow compile against app-crypt/mit-krb5-1.22
sys-auth/sssd uses an unstable API from app-crypt/mit-krb5. Because it
could change at any time, upstream restricts which versions of
app-crypt/mit-krb5 it will work with. So every time there's a new
version of app-crypt/mit-krb5, upstream has to check if the ABI changed
and add the appropriate test for the new version.

Upstream currently cannot automate this for reasons listed in
https://github.com/SSSD/sssd/issues/8068.

Closes: https://bugs.gentoo.org/961637
Signed-off-by: Christopher Byrne <salah.coronya@gmail.com>