sci-physics/espresso: update EAPI 7 -> 8
The upstream project does a wonky. Our buildtype "Gentoo" (eapi 7) is
ignored, and results in RelWithDebInfo using stock cmake flags as the
eclass doesn't erase "foreign" buildtype flags, namely -O3 -DNDEBUG,
which in turn masks a boost assert in a single unittest.  The -9999
ebuild somehow passes this test. Today is not my day for bisecting this.

Bug: https://bugs.gentoo.org/962537
Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>