sci-physics/espresso: make live work, sorta
A couple things of note:
- new option to enable fftw
- all build options renamed with the package name, because cmake
  brainworms :)
- find_package(Python) has changed the case sensitivity for setting the
  executable (PythonInterp was used previously). "yay"
- upstream restricts cython versions and should not. It works fine with
  cython 3.1, and probably will for a while; easier to detect in
  tinderboxing than via "version check in src_configure :P

Most notably, this has two new deps, and violates network-sandbox if not
installed to the system. They aren't packaged yet in Gentoo. This
package "helpfully" builds *and installs* them if needed, assuming you
experiment with FEATURES="-network-sandbox", which is not suitable for a
keyworded package but does serve to verify that the other changes in
this commit are headed in the right direction. They are:

- heFFTe
- HighFive

A task for another day, perhaps.

Signed-off-by: Eli Schwartz <eschwartz@gentoo.org>