dev-util/claude-code: add 1.0.110, drop 1.0.109
The 1.0.110 package no longer contains the vscode .vsix plugin in the
vendor directory. This was not listed as a change in the changelog, but
most things aren't. It may be returning later, it may not, but removing
it from this package since upstream has removed it.

I will readd it later if it gets added back to the package.

Signed-off-by: Jay Faulkner <jayf@gentoo.org>