kernel-install.eclass: verify against SECUREBOOT_SIGN_CERT
The .pcrpkey section of the UKI should not contain a full certificate.
And therefore it is not correct to use it in sbverify. Instead use
the set SECUREBOOT_SIGN_CERT which will contain the certificate that
was used for signing in kernel-build.eclass. For gentoo-kernel-bin
we set this variable to the certificate that was used during build
and is included in the gpkg.

Signed-off-by: Nowa Ammerlaan <nowa@gentoo.org>