sci-libs/blis: Add index64, clean install up
Add a index64 flag as the new consistent flag for ILP64 API.  Unlike
the old index-64bit flag, it installs an ILP64 libblis64 in addition
to the LP64 libblis, therefore avoiding the ABI breakage and matching
the behavior of other packages better.  The naming follows Debian.

While at it, clean up the ebuild a bit.  Install the pkg-config file
to libdir rather than share, and remove the weird Makefile configs
that shouldn't be used by anything.

Closes: https://bugs.gentoo.org/962808
Closes: https://bugs.gentoo.org/962809
Signed-off-by: Michał Górny <mgorny@gentoo.org>