sci-libs/flexiblas: Support installing ILP64 variant
Add an index64 flag to install the ILP64 variant of FlexiBLAS
in addition to the LP64 variant.  It requires the ILP64 variants
of all providers as well.  Note that due to upstream technical
limitations, at the moment this also involves a separate ILP64
sci-libs/lapack version rather than the combined LP64+ILP64 library.

Signed-off-by: Michał Górny <mgorny@gentoo.org>