gui-libs/gtk4-layer-shell: new package, add 1.1.1-r1
Promoted from the existing GURU package, as x11-terms/ghostty-1.2.0
requires it. The revision has been preserved to keep parity with the
version in ::guru, so that once this is merged in ::gentoo and users who
had previously installed it in GURU upgrade their systems, there isn't a
confusing downgrade.

Co-authored-by: unlsycn <unlsycn@unlsycn.com>
Signed-off-by: sin-ack <sin-ack@protonmail.com>
Part-of: https://github.com/gentoo/gentoo/pull/43819
Signed-off-by: Sam James <sam@gentoo.org>