profiles: last rite dev-cpp/rapidyaml
Keeping could potentially have uses to unbundle it in future packages
but it has a tendency to break API (making system's copy problematic)
and is a bit of a pain to maintain and would need extra work to be
"right" (split c4 library, handling python bindings).

Either way, no intention to keep maintaining it myself, so may as well
last rite rather than leave a semi-broken ebuild unless someone wants
to pick it up.

Bug: https://bugs.gentoo.org/963014
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>