qt6-build.eclass: allow lto again in Qt 6.10+ w/ gcc-15.2+
Unclear what fixed what wrt the last failing qtbase test in
bug #955531, but it seems to pass now. Just to be safe, limit
this to recent Qt and gcc.

Tried all of dev-qt/* (except qtwebengine+webview) with lto +
-Werror=odr/mismatch/strict and everything seems to pass.

Note that qtwebengine still filters lto in the ebuild, unlikely
that this is resolved as it was due to some asm usage I recall
(plus we're patching qtwebengine to allow respecting flags at all,
Qt doesn't support users passing lto there).

If other dev-qt/* have issues in the future we should likewise
filter in the ebuild rather have a global one anymore. Formerly
this wasn't about one failing test but tons of them everywhere.

Closes: https://bugs.gentoo.org/955531
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>