sys-libs/glibc: convey locale-gen(8) diagnostics with ewarn()
This commit modifies the behaviour of the run_locale_gen() function
present in the 2.41-r5, 2.41-r6, 2.42, 2.42-r1 and 9999 ebuilds. The
changes are as follows.

Firstly, the responsibility of checking the exit status of locale-gen(8)
and determining whether to die is now that of the function.

Secondly, the standard error stream of locale-gen(8) shall now be
captured. In the event that any lines are captured at all, each shall be
conveyed by the ewarn() function. The purpose of this is to ensure that
diagnostic messages produced by locale-gen(8) are shown not only once
but also a second time, upon emerge exiting. The probability of a user
being aware of any guidance issued by the utility should thus be
markedly increased. Indeed, those that make use of the --quiet-build
option would previously have been oblivious to any such guidance.

Link: https://gitweb.gentoo.org/proj/locale-gen.git/commit/?id=cf37b8be25
Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>