sys-process/parallel: needs ithreads
Unfortunately, parcat really does seem to need ithreads, so depend
on it accordingly. Worse, it doesn't seem to check for failure in all
cases either if it is missing.

Note: if you're hitting a conflict caused by this, you need to set
PERL_FEATURES globally or in package.use with */*, not just for paralllel.

Closes: https://bugs.gentoo.org/940683
Signed-off-by: Sam James <sam@gentoo.org>