app-admin/syslog-summary: Change 'string.' calls
An update in Python removed a bunch of functions from the string module,
including the 'atoi', 'split', and 'rstrip' functions, which
syslog-summary uses, causing it to throw an error. The added patch
changes these function calls to the appropriate format which should
prevent these errors, and removes the 'string' library from the imports.

Closes: https://bugs.gentoo.org/962893
Signed-off-by: Vedant Sule <sulevedant@gmail.com>