dev-qt/qtwebengine: update 6.10 patchset
6.10-4 -> 6.10-[56]:
  Removed:
  - perfetto-pythonpath.patch (upstreamed)

Technically perfetto patch is only in the 6.10.0 branch and not 6.10
yet given the qtwebengine submodule has only been updated in one branch
but it'll most likely be soon or at least by the time of 6.10.1 (that
patch is only for a rare edge case either way).

6.10-5 (not used by any ebuild yet), is the same as 6.10-6 except
for a minor rebase needed due to a change that has made it in the
6.10 branch but not 6.10.0 and likely won't before release (so -5
will likely be used for 6.10.0-rc if nothing else comes up).

Also drop QTBUG-139424.patch from PATCHES, fixed upstream.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>