dev-lang/deno-bin: new package, add 2.5.4
Will (likely) be newly needed by net-misc/yt-dlp to solve YouTube's
JS challenges among other future things, and also been requested
in bug #927139.

I've formerly tried to package the source version, but do not think
I want to be dealing with that after closer inspection. Needs a huge
amount of crates (936+ vendor-filtered, otherwise 1600+) for one,
but the biggest issue is the V8 crate that tries to download various
tools mid-build (or a prebuilt V8 if do not set V8_FROM_SOURCE=1)
and uses Gn to build. Did manage to prevent clang+gn+ninja without
patching but version 2.5.x has an extra hurdle with the rust
toolchain that it downloads and no easy switch to disable it. May
be possible to patch + copy what the chromium ebuild does but it's
all sort of fragile and will be a difficult to maintain across
versions of both deno and the system's toolchain. We *could* use
the same downloaded toolchain through SRC_URI, but then it wouldn't
support anything but amd64+glibc (sounds arm64 would be complicated
too, unless use prebuilt V8... and if using prebuilt may as well use
deno-bin).

In case someone else gets the source build to work for 2.5+, using
"deno-bin" to avoid name clashing with a future package. Either way
we'll likely want to keep -bin given it's a slow build and users
likely won't want to build it just to download youtube videos.

Note that for users that really want to avoid prebuilts or are on an
unsupported profile (e.g. musl), nodejs is technically usable with
yt-dlp but upstream disables it by default due to security concerns
with running untrusted code with it.

Straight to stable and self-keywording for arm64 (without hardware)
given this just dobin a single (~100MB, stripped) file with no tests.

Closes: https://bugs.gentoo.org/927139
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>